# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: GovCon AI Scanner API
  description: |
    Professional REST API for the **GovCon AI Scanner** platform — an enterprise-grade tool
    for searching SAM.gov opportunities, analyzing solicitation documents for bid/no-bid
    decisions, and managing government contracting pipelines.

    ## Authentication

    All protected endpoints require a JWT `Bearer` token in the `Authorization` header:

    ```
    Authorization: Bearer <access_token>
    ```

    Tokens are obtained via `/api/auth/login` or `/api/auth/register` and expire after **15 minutes**.
    Use `/api/auth/refresh` with your `refreshToken` (valid for 7 days) to get a new access token
    without requiring the user to log in again.

    ## Rate Limits

    | Endpoint Group | Limit |
    |----------------|-------|
    | Auth endpoints | 30 req / 15 min per IP |
    | API endpoints  | 200 req / 15 min per IP |
    | Admin endpoints| 200 req / 15 min per IP |

    When a rate limit is exceeded, the API returns **HTTP 429** with a `Retry-After` header.

    ## Error Format

    All errors follow a consistent JSON structure:

    ```json
    {
      "success": false,
      "error": "Human-readable error message",
      "errorCode": "OPTIONAL_MACHINE_CODE"
    }
    ```

    ## Pagination

    List endpoints accept `page` (1-based) and `limit` query parameters and return:

    ```json
    {
      "pagination": {
        "page": 1,
        "limit": 20,
        "total": 150,
...
      }
    }
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Invalid credentials."
        errorCode:
          type: string
          nullable: true
          example: "MISSING_API_KEY"
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        pages:
          type: integer
          example: 8
    User:
      type: object
      properties:
        id:
          type: string
          example: "64f1a2b3c4d5e6f7a8b9c0d1"
        email:
          type: string
          format: email
          example: "jane.doe@acme.gov"
        name:
          type: string
          example: "Jane Doe"
        company:
          type: string
          example: "Acme Defense Solutions"
        naicsCodes:
          type: array
          items:
            type: string
          example: ["541330", "541512", "336411"]
        role:
          type: string
          enum: [user, admin]
          example: "user"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T09:30:00.000Z"
...
paths:
...
  /api/admin/dashboard:
    get:
      tags: [Admin – Analytics]
      summary: Platform KPI dashboard
      description: Returns key metrics — total users, active users, saved opportunities, email subscribers, and more. **Requires admin role.**
      operationId: adminDashboard
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard metrics
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  metrics:
                    type: object
                    properties:
                      users:
                        type: object
                        properties:
                          total:
                            type: integer
                            example: 342
                          active:
                            type: integer
                            example: 318
                          newThisMonth:
                            type: integer
                            example: 47
                          admins:
                            type: integer
                            example: 2
                      opportunities:
                        type: object
                        properties:
                          cached:
                            type: integer
                            example: 15420
                          saved:
                            type: integer
                            example: 891
                      email:
                        type: object
                        properties:
                          subscribed:
                            type: integer
                            example: 287
                      generatedAt:
                        type: string
                        format: date-time
        "403":
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ---------------------------------------------------------------------------
  # Admin – Users
  # ---------------------------------------------------------------------------
  /api/admin/users:
    get:
      tags: [Admin – Users]
      summary: List all users (paginated)
      operationId: adminListUsers
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
            maximum: 100
        - in: query
          name: search
          schema:
            type: string
          description: Search by email, name, or company
        - in: query
          name: role
          schema:
            type: string
            enum: [user, admin]
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
  /api/admin/users/{id}:
    get:
      tags: [Admin – Users]
      summary: Get user details and usage stats
      operationId: adminGetUser
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: "#/components/schemas/User"
                  emailPreferences:
                    type: object
                  stats:
                    type: object
                    properties:
                      savedOpportunities:
                        type: integer
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags: [Admin – Users]
      summary: Update user (admin override)
      operationId: adminUpdateUser
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                company:
                  type: string
                naicsCodes:
                  type: array
                  items:
                    type: string
                role:
                  type: string
                  enum: [user, admin]
                isActive:
                  type: boolean
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  user:
                    $ref: "#/components/schemas/User"
    delete:
      tags: [Admin – Users]
      summary: Deactivate a user account
      operationId: adminDeactivateUser
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User deactivated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "User deactivated."
  /api/admin/users/{id}/activity:
    get:
      tags: [Admin – Users]
      summary: Get user activity log (saved opportunities)
      operationId: adminGetUserActivity
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Activity log
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  activity:
                    type: array
                    items:
                      $ref: "#/components/schemas/Opportunity"
                  pagination:
                    $ref: "#/components/schemas/Pagination"

  # ---------------------------------------------------------------------------
  # Admin – Analytics
  # ---------------------------------------------------------------------------
  /api/admin/analytics/usage:
    get:
      tags: [Admin – Analytics]
      summary: Daily user registration trends
      operationId: adminAnalyticsUsage
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: days
          schema:
            type: integer
            default: 30
            maximum: 365
      responses:
        "200":
          description: Daily registration counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  period:
                    type: string
                    example: "30d"
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        date:
                          type: string
                          example: "2024-01-15"
                        registrations:
                          type: integer
                          example: 12
  /api/admin/analytics/features:
    get:
      tags: [Admin – Analytics]
      summary: Feature adoption statistics
      operationId: adminAnalyticsFeatures
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Feature usage counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  featureUsage:
                    type: object
                    properties:
                      naicsCodeConfiguration:
                        type: integer
                        example: 245
                      emailAlertsEnabled:
                        type: integer
                        example: 189
                      opportunitiesSaved:
                        type: integer
                        example: 891
  /api/admin/analytics/search-trends:
    get:
      tags: [Admin – Analytics]
      summary: Popular NAICS codes and search trends
      operationId: adminSearchTrends
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Search trend data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  naicsCodesByUsers:
                    type: array
                    items:
                      type: object
                      properties:
                        naicsCode:
                          type: string
                          example: "541512"
                        userCount:
                          type: integer
                          example: 87
                  savedOpportunityTrends:
                    type: array
                    items:
                      type: object
                      properties:
                        naicsCode:
                          type: string
                        totalSaves:
                          type: integer
                        uniqueOpportunities:
                          type: integer
  /api/admin/analytics/export:
    get:
      tags: [Admin – Analytics]
      summary: Export user data as CSV or JSON
      operationId: adminExport
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: format
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        "200":
          description: Exported data file
          content:
            application/json:
              schema:
                type: object
            text/csv:
              schema:
                type: string

  # ---------------------------------------------------------------------------
  # Admin – Subscriptions
  # ---------------------------------------------------------------------------
  /api/admin/subscriptions:
    get:
      tags: [Admin – Subscriptions]
      summary: List all subscriptions
      operationId: adminListSubscriptions
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Subscription list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  subscriptions:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/Pagination"
  /api/admin/subscriptions/{userId}:
    get:
      tags: [Admin – Subscriptions]
      summary: Get subscription details for a user
      operationId: adminGetSubscription
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Subscription details
  /api/admin/subscriptions/{userId}/extend-trial:
    post:
      tags: [Admin – Subscriptions]
      summary: Extend trial period for a user
      operationId: adminExtendTrial
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                days:
                  type: integer
                  default: 30
                  example: 15
                  description: Number of additional trial days to grant
                note:
                  type: string
                  example: "Extended for sales demo call on 2024-02-01"
      responses:
        "200":
          description: Trial extended
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Trial extended by 15 days for jane@example.com."
  /api/admin/subscriptions/{userId}/issue-refund:
    post:
      tags: [Admin – Subscriptions]
      summary: Record a refund (process in Stripe dashboard)
      operationId: adminIssueRefund
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  example: 79.00
                reason:
                  type: string
                  example: "Customer dissatisfied with trial experience"
      responses:
        "200":
          description: Refund recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string

  # ---------------------------------------------------------------------------
  # Admin – Health
  # ---------------------------------------------------------------------------
...
  /api/admin/logs:
    get:
      tags: [Admin – Health]
      summary: System logs (stub — use Datadog for production logs)
      operationId: adminLogs
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: level
          schema:
            type: string
            enum: [error, warn, info, debug]
      responses:
        "200":
          description: Log entries (see Datadog for full logs)
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                  logs:
                    type: array
                    items:
                      type: object